pyproject.toml
[tool.poetry]
name = "py-photobooth"
version = "0.1.0"
description = "A small photobooth application for capturing and processing photos."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
keywords = ["photobooth", "camera", "photo", "opencv", "python"]
homepage = "https://example.com/py-photobooth"
repository = "https://example.com/py-photobooth.git"
documentation = "https://example.com/py-photobooth/docs"

[tool.poetry.packages]
# include the package directory (adjust if your package name differs)
py_photobooth = { format = "py", include = "py_photobooth" }

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24"
pillow = "^9.0"
opencv-python = "^4.7"
click = "^8.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^24.0"
isort = "^5.12"
flake8 = "^6.0"
mypy = "^1.9"

[tool.poetry.scripts]
photobooth = "py_photobooth.cli:main"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"